# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/subscriptions:
    get:
      tags:
        - Subscriptions
      summary: List Subscriptions
      description: Get a paginated list of Subscriptions with optional filtering and sorting
      operationId: listSubscriptions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Subscriptions
      summary: Create Subscription
      description: Create a new Subscription
      operationId: createSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/subscription_billing_runs:
    get:
      tags:
        - SubscriptionBillingRuns
      summary: List SubscriptionBillingRuns
      description: Get a paginated list of SubscriptionBillingRuns with optional filtering and sorting
      operationId: listSubscriptionBillingRuns
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionBillingRunList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SubscriptionBillingRuns
      summary: Create SubscriptionBillingRun
      description: Create a new SubscriptionBillingRun
      operationId: createSubscriptionBillingRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionBillingRunRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionBillingRun'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Subscription:
      type: object
      required:
        - id
        - subscription_number
        - customer_id
        - plan_id
        - status
        - started_at
        - current_period_start
        - current_period_end
        - next_billing_date
        - currency
        - metadata
      properties:
        id:
          type: string
          format: uuid
        subscription_number:
          type: string
          maxLength: 40
        customer_id:
          type: string
          format: uuid
        plan_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        started_at:
          type: string
          format: date
        current_period_start:
          type: string
          format: date
        current_period_end:
          type: string
          format: date
        next_billing_date:
          type: string
          format: date
        currency:
          type: string
          maxLength: 3
        cancelled_at:
          nullable: true
          type: string
          format: date
        ended_at:
          nullable: true
          type: string
          format: date
        metadata:
          type: object
...
    CreateSubscriptionRequest:
      type: object
      required:
        - subscription_number
        - customer_id
        - plan_id
        - status
        - started_at
        - current_period_start
        - current_period_end
        - next_billing_date
        - currency
        - metadata
      properties:
        subscription_number:
          type: string
          maxLength: 40
        customer_id:
          type: string
          format: uuid
        plan_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        started_at:
          type: string
          format: date
        current_period_start:
          type: string
          format: date
        current_period_end:
          type: string
          format: date
        next_billing_date:
          type: string
          format: date
        currency:
          type: string
          maxLength: 3
        cancelled_at:
          nullable: true
          type: string
          format: date
        ended_at:
          nullable: true
          type: string
          format: date
        metadata:
          type: object
...
    SubscriptionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Subscription'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    SubscriptionBillingRun:
      type: object
      required:
        - id
        - subscription_id
        - period_start
        - period_end
        - due_date
        - grand_total
        - status
        - idempotency_key
        - metadata
      properties:
        id:
          type: string
          format: uuid
        subscription_id:
          type: string
          format: uuid
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        due_date:
          type: string
          format: date
        grand_total:
          type: string
        status:
          type: string
        invoice_id:
          nullable: true
          type: string
          format: uuid
        idempotency_key:
          type: string
          maxLength: 120
        attempted_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateSubscriptionBillingRunRequest:
      type: object
      required:
        - subscription_id
        - period_start
        - period_end
        - due_date
        - grand_total
        - status
        - idempotency_key
        - metadata
      properties:
        subscription_id:
          type: string
          format: uuid
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        due_date:
          type: string
          format: date
        grand_total:
          type: string
        status:
          type: string
        invoice_id:
          nullable: true
          type: string
          format: uuid
        idempotency_key:
          type: string
          maxLength: 120
        attempted_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    SubscriptionBillingRunList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionBillingRun'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []