# src/presentation/http/subscription_billing_run_handler.rs
//! SubscriptionBillingRun REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc, NaiveDate};
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{SubscriptionBillingRunService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateSubscriptionBillingRunDto, UpdateSubscriptionBillingRunDto, PatchSubscriptionBillingRunDto, SubscriptionBillingRunResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum SubscriptionBillingRunError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for SubscriptionBillingRunError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for SubscriptionBillingRunError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "SUBSCRIPTIONBILLINGRUN_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "SUBSCRIPTIONBILLINGRUN_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "SUBSCRIPTIONBILLINGRUN_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "SUBSCRIPTIONBILLINGRUN_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for SubscriptionBillingRun.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /subscription_billing_runs | List with pagination |
/// | POST | /subscription_billing_runs | Create new |
/// | GET | /subscription_billing_runs/:id | Get by ID |
/// | PUT | /subscription_billing_runs/:id | Full update |
/// | PATCH | /subscription_billing_runs/:id | Partial update |
/// | DELETE | /subscription_billing_runs/:id | Soft delete |
/// | POST | /subscription_billing_runs/bulk | Bulk create |
/// | POST | /subscription_billing_runs/upsert | Upsert |
/// | GET | /subscription_billing_runs/trash | List deleted |
/// | POST | /subscription_billing_runs/:id/restore | Restore |
/// | DELETE | /subscription_billing_runs/empty | Empty trash |
/// | GET | /subscription_billing_runs/:id/deleted | Get deleted by ID |
/// | DELETE | /subscription_billing_runs/trash/:id | Permanent delete from trash |
/// | GET | /subscription_billing_runs/count | Count active entities |
/// | GET | /subscription_billing_runs/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(SubscriptionBillingRunService::with_repository(repository));
/// let router = create_subscription_billing_run_routes(service);
/// ```
pub fn create_subscription_billing_run_routes(service: Arc<SubscriptionBillingRunService>) -> Router {
    BackboneCrudHandler::<SubscriptionBillingRunService, SubscriptionBillingRun, CreateSubscriptionBillingRunDto, UpdateSubscriptionBillingRunDto, SubscriptionBillingRunResponseDto>::routes(
        service,
        "/subscription_billing_runs",
    )
}

/// Create Axum router with only the read (GET) endpoints for SubscriptionBillingRun.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_subscription_billing_run_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_subscription_billing_run_read_routes(service: Arc<SubscriptionBillingRunService>) -> Router {
    BackboneCrudHandler::<SubscriptionBillingRunService, SubscriptionBillingRun, CreateSubscriptionBillingRunDto, UpdateSubscriptionBillingRunDto, SubscriptionBillingRunResponseDto>::read_routes(
        service,
        "/subscription_billing_runs",
    )
}

/// Create Axum router with only the write (mutation) endpoints for SubscriptionBillingRun.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_subscription_billing_run_write_routes(service: Arc<SubscriptionBillingRunService>) -> Router {
    BackboneCrudHandler::<SubscriptionBillingRunService, SubscriptionBillingRun, CreateSubscriptionBillingRunDto, UpdateSubscriptionBillingRunDto, SubscriptionBillingRunResponseDto>::write_routes(
        service,
        "/subscription_billing_runs",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_subscription_billing_run_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<SubscriptionBillingRunService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_subscription_billing_run_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}